requirements.txt
# Core dependencies for Soros Investment Hypothesis Validator

# === AWS & OpenSearch ===
boto3>=1.40.0
botocore>=1.40.0
opensearch-py>=3.0.0
requests-aws4auth>=1.2.3

# === LangChain & LangGraph ===
langchain>=0.3.0
langchain-community>=0.3.0
langchain-aws>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# === OpenAI (for LLM analysis) ===
openai>=1.0.0

# === Data Processing ===
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0

# === Document Parsing ===
beautifulsoup4>=4.12.0
lxml>=4.9.0

# === Environment & Config ===
python-dotenv>=1.0.0

# === Text Processing ===
tiktoken>=0.5.0

# === Testing ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
